+++ client/src/io/worxbend/gitea4s/api/ReposApi.scala
@@ -107,6 +107,12 @@ trait ReposApi:
 
   def assignees(owner: String, repo: String): IO[GiteaError, Chunk[User]]
 
+  def reviewers(owner: String, repo: String): IO[GiteaError, Chunk[User]]
+
+  def stargazers(owner: String, repo: String): ZStream[Any, GiteaError, User]
+
+  def watchers(owner: String, repo: String): ZStream[Any, GiteaError, User]
+
   def collaborators(owner: String, repo: String): ZStream[Any, GiteaError, User]
 
   def isCollaborator(owner: String, repo: String, collaborator: String): IO[GiteaError, Boolean]